data-extractor-webhook/requirements.txt
flask
requests
pdfplumber
pdf2image
reportlab
supabase
python-dotenv
openai
